Refactored some names, removed nullable: false
