+++ src/constants/faction.consts.ts
