+++ src/constants/loadout.consts.ts
