+++ src/constants/world.consts.ts
