+++ src/reports/aggregate/common/combatstats.embed.ts
@@ -1,33 +1,28 @@
 import {Column} from "typeorm";
 
-export class FactionCombat {
+export default class CombatStats{
     @Column({
         type: "number",
-        nullable: false
     })
     kills: number;
 
     @Column({
         type: "number",
-        nullable: false
     })
     deaths: number;
 
     @Column({
         type: "number",
-        nullable: false
     })
     teamKills: number;
 
     @Column({
         type: "number",
-        nullable: false
     })
     suicides: number;
 
     @Column({
         type: "number",
-        nullable: false
     })
     headshots: number;
 }
