+++ src/reports/aggregate/global/outfit.entity.ts
@@ -1,57 +1,50 @@
 import {Column, ObjectIdColumn, Entity, Index, ObjectID} from "typeorm";
-import {World, worldArray} from '../../../constants/world';
+import {World, worldArray} from '../../../constants/world.consts';
 
 @Entity()
 @Index(["outfit", "world"], {unique: true})
-export class OutfitEntity {
+export default class Outfit {
     @ObjectIdColumn()
     _id: ObjectID;
 
     @Column({
         type: "string",
-        nullable: false
     })
     outfit: string;
 
     @Column({
         type: "enum",
         enum: worldArray,
-        nullable: false
     })
     world: World;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     kills: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     deaths: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     teamKills: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     suicides: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     headshots: number;
 }
