+++ src/reports/aggregate/instance/class.entity.ts
@@ -1,57 +1,50 @@
 import {Column, ObjectIdColumn, Entity, Index, ObjectID} from "typeorm";
-import {Loadout, loadoutArray} from '../../../constants/loadout';
+import {Loadout, loadoutArray} from '../../../constants/loadout.consts';
 
 @Entity()
 @Index(["instance", "class"], {unique: true})
-export class ClassEntity {
+export default class Class {
     @ObjectIdColumn()
     _id: ObjectID;
 
     @Column({
         type: "string",
-        nullable: false
     })
     instance: string;
 
     @Column({
         type: "enum",
         enum: loadoutArray,
-        nullable: false
     })
     class: Loadout; // Subject to change to a PlayerInterface
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     kills: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     deaths: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     teamKills: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     suicides: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     headshots: number;
 }
