+++ src/reports/aggregate/instance/weapon.entity.ts
@@ -2,47 +2,41 @@ import {Column, ObjectIdColumn, Entity, Index, ObjectID} from "typeorm";
 
 @Entity()
 @Index(["instance", "weapon"], {unique: true})
-export class WeaponEntity {
+export default class Weapon {
     @ObjectIdColumn()
     _id: ObjectID;
 
     @Column({
         type: "string",
-        nullable: false
     })
     instance: string;
 
     @Column({
         type: "number",
-        nullable: false
     })
     weapon: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     kills: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     teamKills: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     suicides: number;
 
     @Column({
         type: "number",
         default: 0,
-        nullable: false
     })
     headshots: number;
 }
