+++ src/reports/instance/facilitycontrol.entity.ts
@@ -1,41 +1,36 @@
 import {Column, ObjectIdColumn, Entity, Index, ObjectID} from "typeorm";
-import {Faction, factionArray} from '../../constants/faction';
+import {Faction, factionArray} from '../../constants/faction.consts';
 
 @Entity()
 @Index(["instance", "facility", "timestamp"], {unique: true})
-export class FacilityControlEntity {
+export default class FacilityControl {
     @ObjectIdColumn()
     _id: ObjectID;
 
     @Column({
         type: "string",
-        nullable: false
     })
     instance: string;
 
     @Column({
         type: "number",
-        nullable: false
     })
     facility: number;
 
     @Column({
         type: "date",
-        nullable: false
     })
     timestamp: Date;
 
     @Column({
         type: "enum",
         enum: factionArray,
-        nullable: false
     })
     oldFaction: Faction;
 
     @Column({
         type: "enum",
         enum: factionArray,
-        nullable: false
     })
     newFaction: Faction;
 
@@ -47,7 +42,6 @@ export class FacilityControlEntity {
 
     @Column({
         type: "boolean",
-        nullable: false
     })
     isDefence: boolean;
